backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
scipy==1.11.4
scikit-image==0.22.0
pycryptodome==3.19.0
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
PyWavelets==1.5.0
click==8.1.7
colorama==0.4.6
reportlab==4.0.7
